+++ RELEASES.md
@@ -6,6 +6,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-library: updated plead templates to support sql backend, with ID being keys
 * service-templates: deployment of prov-template-library as a service
 * service-templates: submission of csv with json results and reconstructed csv_log
+* service-templates: visualising prov for a template instantiation
 
 `2.0.5`
 * prov-template-library: cleaned up pom.xml